pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.14",
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gevent>=25.5.1",
    "google-genai>=1.25.0",
    "gunicorn>=23.0.0",
    "openai[aiohttp]>=1.95.1",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-pptx>=1.0.2",
    "replit>=4.1.2",
    "reportlab>=4.4.3",
    "werkzeug>=3.1.3",
    "sqlalchemy>=2.0.41",
    "flask-compress>=1.23",
]
